Streamline correlation views
