+++ packages/column-views/stories/correlation-chart/utils.ts
@@ -1,8 +1,11 @@
 import { useArgs } from "storybook/preview-api";
 import { useCallback } from "react";
-import { parseLineFromString, stringifyLine, Identifier } from "../../src";
 import type { ColumnHeaderProps } from "../../src";
+import { Identifier, parseLineFromString, stringifyLine } from "../../src";
 import h from "./stories.module.sass";
+import { fetchUnits, useMacrostratFetch } from "@macrostrat/data-provider";
+import { useAsyncMemo } from "@macrostrat/ui-components";
+import { useCorrelationMapStore } from "@macrostrat/map-views";
 
 export function useCorrelationLine() {
   const [{ focusedLine, selectedUnit }, updateArgs] = useArgs();
@@ -38,4 +41,20 @@ export function CorrelationColumnHeader({
     " ",
     h(Identifier, { id: columnID, className: "column-id" }),
   ]);
+}
+
+export function useCorrelationChartUnits() {
+  const fetch = useMacrostratFetch();
+
+  // Sync focused columns with map
+  const focusedColumns = useCorrelationMapStore(
+    (state) => state.focusedColumns,
+  );
+
+  const colIDs = focusedColumns.map((col) => col.properties.col_id);
+
+  return useAsyncMemo(async () => {
+    const col_ids = focusedColumns.map((col) => col.properties.col_id);
+    return await fetchUnits(col_ids, fetch);
+  }, [colIDs.join(",")]);
 }
